Moved fixtures from Kovats test to subfolder
